+++ diplomacy/src/diplomacy/aop/InwardEdge.scala
@@ -2,7 +2,7 @@ package diplomacy.aop
 
 import chipsalliance.rocketchip.config.Parameters
 import chisel3.Data
-import diplomacy.OutwardNode
+import diplomacy.nodes.OutwardNode
 
 /** Contains information about an inward edge of a node */
 case class InwardEdge[Bundle <: Data, EdgeInParams](
