+++ diplomacy/src/diplomacy/aop/OutwardEdge.scala
@@ -2,7 +2,7 @@ package diplomacy.aop
 
 import chipsalliance.rocketchip.config.Parameters
 import chisel3.Data
-import diplomacy.InwardNode
+import diplomacy.nodes.InwardNode
 
 /** Contains information about an outward edge of a node */
 case class OutwardEdge[Bundle <: Data, EdgeOutParams](
