+++ diplomacy/src/diplomacy/bundlebridge/BundleBridgeEphemeralNode.scala
@@ -1,0 +1,6 @@
+package diplomacy.bundlebridge
+import chisel3.Data
+import diplomacy.nodes.EphemeralNode
+
+case class BundleBridgeEphemeralNode[T <: Data]()(implicit valName: sourcecode.Name)
+    extends EphemeralNode(new BundleBridgeImp[T])()
