+++ diplomacy/src/diplomacy/bundlebridge/BundleBridgeIdentityNode.scala
@@ -1,0 +1,6 @@
+package diplomacy.bundlebridge
+import chisel3.Data
+import diplomacy.nodes.IdentityNode
+
+case class BundleBridgeIdentityNode[T <: Data]()(implicit valName: sourcecode.Name)
+    extends IdentityNode(new BundleBridgeImp[T])()
