+++ diplomacy/src/diplomacy/nodes/package.scala
@@ -1,0 +1,8 @@
+package diplomacy
+
+import chisel3.Data
+
+package object nodes {
+  type SimpleNodeHandle[D, U, E, B <: Chisel.Data] = NodeHandle[D, U, E, B, D, U, E, B]
+  type AnyMixedNode = MixedNode[_, _, _, _ <: Data, _, _, _, _ <: Data]
+}
